Move API-related funcs to api.go
